Time_Trigger/requirements.txt
azure-functions
azure-storage-blob==12.0.0
wget==3.2